+++ src/RestApi/Location/OrganizationTag/LocationOrganizationTag.ts
@@ -1,0 +1,9 @@
+export interface AssignOrganizationTagToLocation {
+	locationUid: string;
+	tagUid: string;
+}
+
+export interface UnassignOrganizationTagFromLocation {
+	locationUid: string;
+	tagUid: string;
+}