+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Array `supportedResolutions` into Device entity
 - Bulk Operations
 - Location `create`, `list` (multiple), `get` (single), `update` and `delete`
+- Organization tag to location `assign` and `unassign`
 
 ## [1.7.1] - 2021-12-17
 ### Fixed